WIP: Vectorising training batches + training loop
To get GPU speedup we need to train on batched data, however currently it uses a batch size of 1 in a for loop (not ideal at all). Since actions and targets are variable length tensors we need to figure out correct padding etc and stack them to create a e.g. 1024 batch tensor tuple.