Big update: Replay Buffer + Training revised
Went through and completely changed how the replay buffer saves replays so that training examples are vectorized/batched and hence can be done in parallel. Batches are now typically (N,K+1) dimensions (except for the observation tensor and policy targets).
Fixed up a lot of bugs in the training routine so now it runs end to end.
Had an initial play around with loss functions, using MSE for simplicity for now.