WIP: remote storage + test simulation
Added a way to render every nth game so that progress can be monitored.
Began working on a way to make the network storage buffer remote. Ray can't serialise tensorflow models so it will need to save and retrieve the weight tensors. We have 2 options, save to memory as something like a protobuf or read/write keras saved model weights from disk.