Updated loss function - more typical loss curves now
Made sure to only divide by non-masked examples in the reduce_mean reductions. This now shows that the network is converging to something close to the base entropy (i.e. KL is pretty close to 0) and hence our network is doing a very good job of tracking the targets (although there seems to be an offset in the means). Reward overall still sucks and it fails to learn meaningful policies...